+++ R/FileSurveyInput.R
@@ -48,7 +48,10 @@ FileSurveyInput <- function(id, label = "Upload data (csv/xlsx/sav/sas7bdat/dta)
   tagList(
     fileInput(ns("file"), label),
     uiOutput(ns("factor")),
-    uiOutput(ns("survey"))
+    uiOutput(ns("survey")),
+    uiOutput(ns("subset_check")),
+    uiOutput(ns("subset_var")),
+    uiOutput(ns("subset_val"))
   )
 }
 
@@ -163,7 +166,7 @@ FileSurvey <- function(input, output, session, nfactor.limit = 20) {
     #except_vars <- names(nclass)[ nclass== 1 | nclass >= 10]
     add_vars <- names(nclass)[nclass >= 2 &  nclass <= 5]
     #factor_vars_ini <- union(factor_vars, add_vars)
-    return(list(data = out, conti_original = conti_vars, factor_adds_list = names(nclass)[nclass <= nfactor.limit], factor_adds = add_vars, ref = ref, naomit = naomit))
+    return(list(data = out, factor_original = factor_vars, conti_original = conti_vars, factor_adds_list = names(nclass)[nclass <= nfactor.limit], factor_adds = add_vars, ref = ref, naomit = naomit))
   })
 
 
@@ -211,6 +214,38 @@ FileSurvey <- function(input, output, session, nfactor.limit = 20) {
                     selected = selected.weight.final)
         )
       })
+  })
+
+  observeEvent(c(data()$factor_original, input$factor_vname, input$repeated_vname, input$cluster_vname, input$strata_vname, input$weights_vname), {
+    output$subset_check <- renderUI({
+      checkboxInput(session$ns("check_subset"), "Subset data")
+    })
+  })
+
+  observeEvent(input$check_subset, {
+    output$subset_var <- renderUI({
+      req(input$check_subset == T)
+      factor_subset <- setdiff(c(data()$factor_original, input$factor_vname), c(input$repeated_vname, input$cluster_vname, input$strata_vname, input$weights_vname))
+
+      validate(
+        need(length(factor_subset) > 0 , "No factor variable for subsetting")
+      )
+
+      tagList(
+        selectInput(session$ns("var_subset"), "Subset variable",
+                    choices = factor_subset, multiple = F,
+                    selected = factor_subset[1])
+      )
+    })
+
+    output$subset_val <- renderUI({
+      req(input$check_subset == T)
+      req(input$var_subset)
+      varlevel <- levels(as.factor(data()$data[[input$var_subset]]))
+      selectInput(session$ns("val_subset"), "Subset value",
+                  choices = varlevel, multiple = F,
+                  selected = varlevel[1])
+    })
   })
 
 
@@ -262,6 +297,23 @@ FileSurvey <- function(input, output, session, nfactor.limit = 20) {
       w <- which(ref[["name.new"]] == vn)
       out.label[variable == vn, var_label := ref[["name.old"]][w]]
     }
+
+    if (!is.null(input$check_subset)){
+      if (input$check_subset){
+        validate(
+          need(length(input$var_subset) > 0 , "No factor variable for subsetting")
+        )
+        out <- out[get(input$var_subset) == input$val_subset]
+        surveydata <- survey::svydesign(id = cluster.survey, strata = strata.survey, weights = weights.survey, data = out)
+        var.factor <- c(data()$factor_original, input$factor_vname)
+        out[, (var.factor) := lapply(.SD, factor), .SDcols = var.factor]
+        out.label2 <- mk.lev(out)[, c("variable", "class", "level")]
+        data.table::setkey(out.label, "variable", "class", "level")
+        data.table::setkey(out.label2, "variable", "class", "level")
+        out.label <- out.label[out.label2]
+      }
+    }
+
     return(list(data = out, label = out.label, naomit = data()$naomit, survey = surveydata))
   })